+++ src/main/java/ballistix/client/screen/ScreenFireControlRadar.java
@@ -84,10 +84,10 @@ public class ScreenFireControlRadar extends GenericScreen<ContainerFireControlRa
             width = font.width(radar);
 
             if (width > 100) {
-                scale = 100.0F / (float) width;
+                scale = 100.0F / width;
             }
 
-            float remHeight = (height - (float) height * scale) / 2.0F;
+            float remHeight = (height - height * scale) / 2.0F;
 
             graphics.pose().pushPose();
